reorganize nuweb cleaning tasks
nuweb:clobber still removes all generated files, but is not a prerequisite for the global clobber.
nuweb:clean does not remove sentinel (.ws) files used for nuweb-building.
an intermediate nuweb task clean_more does clean and removes .ws files; this is a prerequisite for the global clobber.
nuweb:tangle is now a prerequisite for gem:package